Update Movement::MovementArray#forward and add comments
While not beautiful, #forward now takes two arguments: the
integer/range/array and a reference to the piece itself. This allows it
to check which team it is on and decided which direction to move.

Debating on whether to move that logic into the piece itself, or move it
further up to game. Will decided later.