Update MovementArray and Piece Object
MovementArray now accepts a piece object instead of a position, allowing
it to more accurately determine the possible positions a piece can move
to. Previously, it could only give an array of coordinates relative to
the coordinate given without consideration for the board itself or
blocking pieces.

This update makes the #get_possible_moves unnecessary, as now
MovementArray can generate all normal, legal moves that a piece can
make.

Piece objects now have a @board instance variable, which points to the
board in which they are placed. Pieces also have a #add_to_board method,
which calls the Board's #place method, giving itself and its starting
position, and updating its @board variable with the board provided.