Add Gamestate class and check/checkmate methods
Create a Gamestate class, which keeps a log of all pieces and their
positions in each state. Using it, a board can rewind itself to a
previous state.

in_check? and checkmate? movements are added to Movement module.
Checkmate? utilizes the gamestate class to rewind the board each time it
makes a move, allowing it to try out every possible move to get out of
check.

May consider creating a copy of board instead to minimize any side
effects to the real board.