Add ChessNotation Module
ChessNotation converts to and from notation used in chess.

Public interface consits of ChessNotation#from_notation and
ChessNotation#to_notation.

from_notation takes a string and returns a hash with information
regarding the move, primarily the piece involved and where it moved, as
well as supplemental information regarding special moves such as
castling, promotions, etc. (still need to fine tune).

to_notation takes two positions and a board as arguments. The first
position is the location of the piece to move and the second position is
its destination. It returns a string that represents a move in chess
notation.