Add Map &  ColorMap classes and Highlighting mod.
The map class is responsible for creating an association between a
string and an array and displaying the string with the contents of the
array injected into it.

An array, a string and a key are passed into the map on initialization.
The string is then parsed for the key character, and each key position
is marked. The array is similarly parsed for all element positions and
the two position sets are stored in a hash as key-value pairs. An
inverted hash is also stored.

When displaying the string, each key-marked position is replaced by the
contents of the related array position (or a value stored at @empty_chr
if content is nil).

ColorMap is a subclass of Map and, in addition to matching array and
string positions, accepts strings representing background colors and
font colors. It creates arrays representing colors for each position in
the string. Using the Highlighting module, it colors in the string
according to the color array before display.