Optimize Movement and related methods
Put Movement module into its own file (movement.rb) and optimize code,
particularly the in_checkmate? method.

in_checkmate? calculated every single move that a set could make, which
was unnecessary. Added another step, where it first checks if the king
can escape check on its own (a maximum of eight checks). If it fails
that check, then it gets all spaces that are between the king and its
attackers (including attacker positions). While going through each ally
piece's moves, it checks if that move enters any of those important
spaces. If not, it doesn't simulate the move.

in_check? has been updated to return its array of found attacking
pieces, or nil if the array is empty. The values are truthy/falsy,
allowing it to be used the same way.

board now sets @moved to true whenever it moves a piece.

Setting @moved on king revealed an error in its special_moves method,
where it would return nil if it had already moved. Added an empty array
to return if moved is true.