Further refine move structure
The previous iteration of move looked like this:

	{ piece1.id => destination_pos1,
  	  piece2.id => destination_pos2 }

While it was lightweight, it required a lot of requests to objects to
get enough information for certain calculations. Furthermore, it was
impossible to undo a move with only the move itself as information.

The new move looks like this:

	{ piece1 => [ current_pos1, destination_pos1 ],
  	  piece2 => [ current_pos2, destination_pos2 ] }

Now, a move can be applied to the board or undone from the board with no
extra information required. This makes calculating check and checkmate
much easier, as creating duplicate boards is no longer necessary.

ex.
	@board.do(move)
        check = **.in_check?(king, @board)
        @board.undo(move)

Now, board also has its own in_check? and checkmate? methods. in_check? takes
a hash and expects either the king or its team. checkmate? only
accepts a team as a parameter. Game has been updated to use board's
methods instead of Movement's.