Turn Move into an object
Turn Move into an object and add methods to make calculations using
moves more intuitive.

Methods include: #destination(piece), which gets the destination pos for
piece, #include?(pos), which goes over ever destination position in move
and checks for any matches, #each(&block), which calls block for every
set of instructions in move.

Also create an EmptyMove class, which stands in for a move when no move
can be found.