Add StateTree, State and Observer classes
A big overhaul of how gamestate data is handled. Instead of relying on
the board to store the state of the game, each state is held in a state
object which is stored as data in a StateTree.

The State object has several hashes with locations of every piece, their
moved status and possible moves. It also stores the status of check and
checkmate at any given time (once they have been calculated).

The StateTree is the gamestate interface that the rest of the game
interacts with. Whenever the current_node is updated, any Observers are
notified and given the StateTree as a parameter so they can update
accordingly.

This change will make it much easier to implement a decision-making
algorithm for a computer player, as any calculations it does looking for
checkmate can be stored and used again later.

One flaw: currently possible moves are generated all at once by the
State class when it is created. However, the en passant move needs to
know about a previous state in order to calculate its eligibility. Need
to find a new way to calculate en passant, or change who is in charge of
initiating the calculation.