Add highlighting for selected piece
CursorMap now automatically highlights a selected piece and its possible
moves. When the selected piece is unselected, all highlights are
removed (via on_backspace).

Added a call to update in before_get_input so in the case that an
invalid move is entered, the window is updated (and selection highlights
removed) to indicate that no piece is currently selected.

Note to self: need to find a way to inject these rules instead of
hard-coding them.