Add Promotion Functionality
The game now checks for a promotion from the move selected by the user
(before the move is validated). A new board is overlaid on the game
board with the player's choices of promotion pieces. When the player
chooses a piece, the piece is set in move's set_promotion_piece
function.

Things to look out for: if the move is altered after the move has been
validated, the gamestate will have already stored a state connected to
it (when checking check). Then, the gamestate will active that stored
state regardless of updates to the @instructions set. You either need to
create a new object or change the #== method.