Fix bug in Promotion Handling
When a pawn had multiple moves that led to promotion, only the first
move was properly deleted. When Movement.return_move checked for
possible moves, it would see the undeleted, non-promotion version of a
move and return that instead of the newly created promotion version.

Now the pawn movement generating function properly deletes any moves
that are replaced by promotion moves.