Add Move Validation Against Self-check
Added a new method to Movement called validate_moves, which takes a set
of moves (all moves for a particular team) and checks that each move
does not lead to self-check or leave king in check and sets the invalid
parameter of any failing moves to true.

Now, pieces only return possible moves that have an invalid parameter of
false.

The game object validates all moves for a team before their turn begins.