Fix validation of moves when King is in check
While the algorithm in Movement#validate_moves properly tagged any
self-checking or non-saving moves, the moves did not retain the :invalid
attribute. This was due to the handling of the king's moves, which were
applied and unapplied to the gamestate, causing moves to be recreated
with new object ids.

While it may be possible to check the king's moves without changing the
gamestate in the future, for now it is easier to return the tagged moves
and then update the gamestate to hold them.

Now, gamestate has an update_moves method which does exactly that.