Fix bug affecting move lists in State object
In State#do, the @pieces hash is copied and the new move is applied to
it before creating a new state. However, the hashes for each piece
inside of @pieces were not copied. These hashes were then merge!ed with
new move_sets, which would be retroactively applied to previous piece
hashes, causing subtle errors in certain situations. Each piece hash is
now individually copied to prevent this behavior.