Game Loading Now Works
More changes to the initialization of certain objects including Game and
Player, specifically to allow non-serialized variables to be set after
the objects are created. This allows ui elements to be defined after the
game object is created and allows InputHandler to be inserted into
Player after it is created.

A new method called init_game_ui in lib/chess.rb takes a game object, creates ui elements
and inserts them into the game. The game now inserts the InputHandler
into player objects in the start method.

One important bug was patched here: the @positions hash of State was not
properly loaded (keys that should be arrays were strings) making it
impossible for board to access pieces at each location. @positions is
generally generated on State initialization anyway, so ignoring
@positions on serialization makes everything work fine.