Update InteractiveScreen and InputHandler
Make InteractiveScreen more flexible by adding @key_map parameters to
set unique behavior for certain key presses. At the same time, update
how InputHandler handles instructions.

InputHandler previously only accepted Procs/lambdas as instructions,
immediately calling them when it the matching key is pressed. However,
the Proc/lambda retains the context in which it was created, making it
very difficult to create an object with unique input handling (since the
instructions would be based on the object creator's context, and not the
object's itself).

Now, InputHandler also accepts strings as instructions. The string is
the name of the object's method to be called, which is retrieved via the
"method" method.