Further update Interactive/InputHandler
On creating a UI element featuring mulitple interactive screens, it
became apparent that using the "method" method on its own was not
sufficient. For example, if an InteractiveScreen is nested within
another InteractiveScreen, the outer screen's method will be retrieved
instead of the inner screens method.

To properly get the method of nested elements, Interactive Windows now
have a "get_action" method. It is normally set to call the method
method, but can be changed to accommodate nesting elements.

In the case of InteractiveScreen, it now checks if the active_region can
respond to the method given, and if so, retrieves active_region's method
isntead.

WindowTemplates now has a working multipage_window template as well.