First working Settings Menu Template
Add settings_menu method in WindowTemplates.

Settings menu features a title, main menu, a "Save" and "Cancel" button,
and sub menus for each setting in the main menu.

Because of the complicated behavior of and interactions of different
elements, the previous method of assigning key presses to actions needed
to be updated. The most important update was simple: allow key maps to
be set or merged after an object is declared. That way, it is easy to
specify which object to call, rather than depending on an unattached
call (via #method(mthd_name)) to reach its proper destination.

Also added the ability to read the @content property and set the
input_to_return attribute from outside the object.