Fix bug in calculating Checkmate
Statetree.checkmate method incorrectly calculated checkmate when the
attacking piece was a knight.

To solve, copy-pasted a portion of the logic from
Movement.validate_moves method into checkmate method, allowing it to
properly calculate important spaces (in case of Knight, only Knight's
current position).

Would like to isolate logic in the future, but don't want to create new
dependencies for the moment.