Fix bugs with Chess AI
Updated minimax algorhithm and fixed bugs with gameplay loop with AI.

Sometimes the AI chooses moves that are deemed "invalid" by the game
object but are perfectly fine. This led to some strange errors. Changing
the valid_move check to see whether or not the move found is empty
solved this.

Added get_valid_moves method to statetree to streamline getting possible
moveset. The statetree now stores the active team and uses that to get
and validate moves.