Update Game Save Logic
Previously, the game object was fully serialized when being saved. Due
to the organization of the code, which has some back-and-forth object
references, serializing this way quickly spiraled out of control and
caused errors.

Now, instead of serializing the game object, only the player objects and
a list of moves are serialized instead. These are then loaded into the
load method, which plays out every move given. The game can then have
its ui initialized and played out normally.