Set Maximum Saves and add Save Overwrites
Now there are a maximum number of save slots, each of which is occupied
by a Save or EmptySave object. When there are no empty slots available
to save, the game prompts the user to overwrite a save or quit without
saving.

This update includes some streamlining for window elements, namely
separating the program's control logic from the window elements, and
instead returning simple inputs from the user, such as the chosen index
number of an item in a list, or a 1 or 0 from a confirmation screen.