Fix bugs with castling calculations
Castling was missing several conditions for calculation and also had a
bug where rank and file were switched in positions for king and rook.

The method now checks first whether the king is in_check or has moved
before calculating castle moves, and also checks for any attacking
pieces on the spaces that the king traverses during castling.