Add linebreak logic for Window class
Window class will now break a string down into lines based on its size.
If characters besides spaces are found at the end of the current string
and beginning of the next string, all characters will be moved to the
next line to prevent word breaks.

The about game method in lib/chess.rb was slightly changed to test this
new logic.