Update WindowTemplates with bg settings
Add default_background_settings and interactive_screen methods to
WindowTemplates, as well as color_set method.

Allow WindowTemplates to get selected color set directly so that
arguments do not have to be passed to each template with color set's
colors. This makes it easier to use templates in different areas of the
program.

Fix some parameter names to better match the structure of
WindowTemplates. For example, instead of :title, pass :title_content.