Fix loss of color on TypingField.on_backspace
TypingField.on_backspace changed the content of the window then called
the update method to redraw the window. Because the standard update
method sets the color attribute on and off during update, that left the
color attribute OFF after backspace was pressed.

Instead of calling update, delete the character with @win.delch instead.

Also, removed the line that reset @input_to_return to "" in
before_get_input, in case a user moved between the typing field and
button and wanted to edit their input again.